+++ data/scripts/galaxy/eclipse_roaming_boss.lua
@@ -1,7 +1,7 @@
 package.path = package.path .. ";data/scripts/lib/?.lua"
 package.path = package.path .. ";data/scripts/?.lua"
 
-local cv_success, cv_news = pcall(require, "cosmicvaultnews")
+local cv_success, cv_news = true, require("cosmicvaultnews")
 
 EclipseRoamingBoss = {}
 local self = EclipseRoamingBoss